Add project variable to limit number of references in network objects
Use of the new project variable `LIMIT_NETWORK_REFERENCES` allows to
get the same behaviour as in the `tn-ro-reduced` branch. It reduces the
number of object references in the network objects to the number set
in the variable (exception: if the variable is set, the
`TransportNetwork` object will contain no references at all).

If the variable is not set, the number of references is not limited
(original behaviour).