fix: update build.gralde file to use it with windows
With windows the default path seperator is a '\'. This leads to an error when running a
git diff: In the diff file the path to the respective files is inserted with a '\\' separator.
And when apllying this diff file the respective files (with '\\' separator) can't be found.
This is changed here so that the '\' is replaced by a '/'.

Also Parameter '--ignore-whitespace' is needed when running a git apply. Otherwise existing
(most likely with linux created) diff-files can't be used with windows.

SVC-1734