feat: update tn-a to PostNAS 7.1
Automatic migration was done using the Gradle task 'migrate-tn-a'.

The existing diff-file was used to apply the already existing manual changes.
Therefore the namespaces used in the diff file were adjusted and when running the
task 'migrate-tn-a' the first time the command to create a new diff was deactivated.

In addition in the newly added groovy scrips attribute names were adjusted and
duplicate relation of attribute to 'AerodomeArea' were deleted.

SVC-1734