feat: update tn-ra to PostNAS 7.1
Automatic migration was done using the Gradle task 'migrate-tn-ra'.

The existing diff-file was used to apply the already existing manual changes.
Therefore the namespaces used in the diff file were adjusted and when running the
task 'migrate-tn-a' the first time the command to create a new diff was deactivated.

Not all existing changes were applied as some of adjusted cells were deleted in the
new auto-alignment. Some previous manual adjustments where updated manually as the
auto-alignment changed and therefore the changes could be applied automatically.

SVC-1734