feat: update tn-ro to PostNAS 7.1
Automatic migration was done using the Gradle task 'migrate-tn-ro'.

The existing diff-file was used to apply the already existing manual changes.
Therefore the namespaces used in the diff file were adjusted and when running the
task 'migrate-ro' the first time the command to create a new diff was deactivated.

Also two filters on feature type level were adjusted in the  existing diff file
(the filters changed due to the update to 7.1) to be able to apply the related changes.

In two Groovy Typ relations a new syntax was added due to the update to 7.1. This syntax
was adjusted to explicitly use codes as strings as the respective attributes are defined
as numbers in the postNAS schema.

SVC-1734