feat: update tn-w to PostNAS 7.1
Automatic migration was done using the Gradle task 'migrate-tn-w'.

The existing diff-file was used to apply the already existing manual changes.
Therefore the namespaces used in the diff file were adjusted and when running the
task 'migrate-tn-w' the first time the command to create a new diff was deactivated.

Additionally one comment in the existing diff-file was adjusted (the comment changed
due to the update to 7.1) to be able to apply the related changes.

In one groovy script a new code block was added with the update to 7.1. This syntax
was adjusted to explicitly use codes as strings as the respective attribute is defined
as numbers in the postNAS schema.

SVC-1734