feat: update cp to PostNAS 7.1
Automatic migration was done using the Gradle task 'migrate-cp'.

The existing diff-file was used to apply the already existing manual changes.
Therefore the namespaces used in the diff file were adjusted and when running the
task 'migrate-cp' the first time the command to create a new diff was deactivated.

No other manual changes were needed and done.

SVC-1734