feat: update au to PostNAS 7.1
Automatic migration was done using the Gradle task 'migrate-au'.

The existing diff-file was used to apply the already existing manual changes.
Therefore the namespaces used in the diff file were adjusted and when running the
task 'migrate-au' the first time the command to create a new diff was deactivated.

Additional manual changes: multiple occurrences of relation to attributes in
'AdministrativeUnit' were deactivated.

SVC-1734