feat: update hy-n to PostNAS 7.1
Automatic migration was done using the Gradle task 'migrate-hy-n'.

The existing diff-file was used to apply the already existing manual changes.
Therefore the namespaces used in the diff file were adjusted and when running the
task 'migrate-hy-n' the first time the command to create a new diff was deactivated.

Additional manual changes:
With the update to 7.1 a new new groovy join function to 'WatercourseSeparatedCrossing'
is added. This script is adjusted to use '_source.p.gml_id.value()' instead of
'_source.p.id.value()'.
Also multiple occurrences of the same relation to attributes in 'WatercourseSeparatedCrossing'
are deactivated.

SVC-1734