feat: update bu-2d to PostNAS 7.1
Automatic migration was done using the Gradle task 'migrate-bu-2d'.

The existing diff-file was used to apply the already existing manual changes.
Therefore the namespaces used in the diff file were adjusted and when running the
task 'migrate-bu-2d' the first time the command to create a new diff was deactivated.

Additional manual changes:
With the new adv alignment two filters changed and the related manual changes to replace
'_source.p.id.value()' with '_source.p.gml_id.value()' could be applied automatically. This
was adjusted manually.

In the new adv alignment the mapping of attributes 'hoehe' and 'uom' changed. Therefore
the existing manual changes (to replace it by an assignBound with value 'm') couldn't be
applied automatically. This was adjusted manually.

Previously, the mapping of attributes 'name' in different feature types were removed.
This is changed here so that the mapping is applied correctly.

Multiple occurrences of identical relation to attributes of 'BuildingPart' were removed.

SVC-1734