feat: add custom type to PostgreSQL schema for Bundesland extent
In the adv-inspire-alignments, the ELU alignment has a special mechanism to
include the extent of a German Bundesland via an additional CSV source file.
Because of how the postnas-inspire-alignments are derived from the
adv-inspire-alignments, this additional source information needs to be
present in the matching alignment.

To avoid having to deal with both a database and a file source, a custom table
is added to the PostgreSQL schema instead and included in the matching
alignment. This table can be added by users to their PostgreSQL source database
to provide the Bundesland extent.

SVC-1734