feat: update mapping au to new INSPIRE schema 5.0
Automatic migration ran successfully, no manual changes were needed.

Additionally the diff-file was adapted manually for the change related to the cell with
id="C8dac57f2-ccc4-4666-997a-6900aedfd82a". Here a value for'AdministrativeUnit.UpperLevelUnit'
is assigned. When running the migration with windows the respective change is added
automatically (and there it is contained in the updated alignment). But when running the
task with Linux it was necessary to add the change manually. There the namepsace in the diff-file
was adapted manually and kept.

SVC-1974