fix: correct date parsing in tn-ra alignment
Previously, 'Date.parse' was used to convert the RLP date format to a date object.
With the newer hale studio version this function is not available any longer. Therefore
'SimpleDateFormat' is used instead.

The change was done in the manual alignment and the gradle task 'migrate-tn-ra' was run.

Additionally a note on the special date handling for RLP date format is added in the md-file
as this was missed before.

WGS-3143